+++ packages/block/src/index.ts
@@ -1,5 +1,5 @@
 export { Block } from './block.js'
-export * from './blockConstructor.js'
+export * from './constructors.js'
 export { type BeaconPayloadJson, executionPayloadFromBeaconPayload } from './from-beacon-payload.js'
 export { BlockHeader } from './header.js'
 export {
