+++ packages/block/test/eip4895block.spec.ts
@@ -10,10 +10,7 @@ import {
 } from '@ethereumjs/util'
 import { assert, describe, it } from 'vitest'
 
-import {
-  createBlockFromBlockData,
-  createBlockFromRLPSerializedBlock,
-} from '../src/blockConstructor.js'
+import { createBlockFromBlockData, createBlockFromRLPSerializedBlock } from '../src/constructors.js'
 import { BlockHeader } from '../src/header.js'
 import { genWithdrawalsTrieRoot } from '../src/helpers.js'
