+++ packages/block/test/from-beacon-payload.spec.ts
@@ -3,7 +3,7 @@ import { loadKZG } from 'kzg-wasm'
 import { assert, beforeAll, describe, it } from 'vitest'
 
 import * as shardingJson from '../../client/test/sim/configs/4844-devnet.json'
-import { createBlockFromBeaconPayloadJson } from '../src/blockConstructor.js'
+import { createBlockFromBeaconPayloadJson } from '../src/constructors.js'
 import { BlockHeader } from '../src/index.js'
 
 import * as payloadKaustinen from './testdata/payload-kaustinen.json'
